# docs/swagger.yaml
definitions:
  message.CasbinInReceive:
    properties:
      CasbinInfos:
        items:
          $ref: '#/definitions/message.CasbinInfo'
        type: array
      ExtendsRoleIds:
        description: 继承角色id
        items:
          type: string
        type: array
      RoleId:
        description: 角色id
        type: string
    type: object
  message.CasbinInfo:
    properties:
      Method:
        type: string
      Path:
        type: string
    type: object
  message.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  message.LoginResp:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        type: string
    type: object
  message.PolicyPathResponse:
    properties:
      Paths:
        items:
          $ref: '#/definitions/message.CasbinInfo'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "2.0"
paths:
  /base/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回包括用户信息,token,过期时间
          schema:
            $ref: '#/definitions/message.LoginResp'
      summary: 用户登录
      tags:
      - Base
  /menu:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/message.LoginResp'
      security:
      - ApiKeyAuth: []
      summary: 获得用户菜单
      tags:
      - Menu
  /role/permission:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 删除角色，改角色被人继承时不能被删除，请先取消继承关系
          schema:
            type: string
      summary: 删除角色
      tags:
      - Casbin
    get:
      consumes:
      - application/json
      parameters:
      - description: 角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 获取权限列表,返回包括casbin详情列表
          schema:
            $ref: '#/definitions/message.PolicyPathResponse'
      summary: 获取权限列表
      tags:
      - Casbin
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 给角色新增权限
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 给角色新增权限
      tags:
      - Casbin
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 更新角色api权限
          schema:
            type: string
      summary: 更新角色api权限
      tags:
      - Casbin
  /role/permission-extends:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 角色id, 继承角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 删除角色继承
          schema:
            type: string
      summary: 删除角色继承
      tags:
      - Casbin
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色id, 继承角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/message.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 给角色新增继承
          schema:
            type: string
      summary: 角色继承
      tags:
      - Casbin
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"